+++ src/write_wat.c
@@ -241,7 +241,7 @@ void web49_wat_print_section_table(web49_io_output_t *out, web49_module_t mod, w
     for (uint64_t i = 0; i < stable.num_entries; i++) {
         web49_type_table_t table = stable.entries[i];
         web49_io_output_fprintf(out, "\n  (table (;%zu;) %" PRIu64, (size_t)i, table.limits.initial);
-        if (table.limits.maximum != UINT64_MAX) {
+        if (table.limits.maximum != UINT32_MAX) {
             web49_io_output_fprintf(out, " %zu", (size_t)table.limits.maximum);
         }
         web49_io_output_write_str(out, " ");
@@ -257,7 +257,7 @@ void web49_wat_print_section_table(web49_io_output_t *out, web49_module_t mod, w
 void web49_wat_print_section_memory(web49_io_output_t *out, web49_module_t mod, web49_section_memory_t smemory) {
     for (uint64_t i = 0; i < smemory.num_entries; i++) {
         web49_type_memory_t mem = smemory.entries[i];
-        if (mem.maximum != UINT64_MAX) {
+        if (mem.maximum != UINT32_MAX) {
             web49_io_output_fprintf(out, "\n  (memory (;%zu;) %zu %zu)", (size_t)i, (size_t)mem.initial, (size_t)mem.maximum);
         } else {
             web49_io_output_fprintf(out, "\n  (memory (;%zu;) %zu)", (size_t)i, (size_t)mem.initial);
